Docserv: Set limit for open files, scale down threads if we can't (#199)
We have had a bunch of issues with Docserv2 erroring out with "OSError:
Too many open files". You can solve that in the machine's config but
it's probably nicer to solve it in the program.

The number of 200 files per thread is a wild guesstimate by looking at
lsof -u USER | wc -l, and may actually be a bit low.